+++ src/main/java/pl/edu/pwr/ksiegarniainternetowa/service/BookService.java
@@ -4,6 +4,6 @@ import pl.edu.pwr.ksiegarniainternetowa.model.entity.BookEntity;
 
 public interface BookService {
 
-    public BookEntity getBookById(Long bookId);
+    public BookEntity getBookById(Integer bookId);
 
 }
